requirements.txt
asgiref==3.3.4
attrs==21.2.0
autopep8==1.5.7
Babel==2.9.1
certifi==2021.5.30
cffi==1.14.5
chardet==4.0.0
coreapi==2.3.3
coreschema==0.0.4
cryptography==3.4.7
defusedxml==0.7.1
Django==3.2.4
django-allauth==0.44.0
django-cors-headers==3.7.0
django-filter==2.4.0
django-money==2.0
django-phonenumber-field==5.2.0
django-rest-auth==0.9.5
django-rest-passwordreset==1.2.0
django2-rest-passwordreset==1.3.5
djangorestframework==3.12.4
drf-yasg==1.20.0
drf-yasg2==1.19.4
idna==2.10
inflection==0.5.1
itypes==1.2.0
Jinja2==3.0.1
jsonschema==3.2.0
Markdown==3.3.4
MarkupSafe==2.0.1
oauthlib==3.1.1
packaging==21.0
Pillow==8.2.0
psycopg2==2.9.1
psycopg2-binary==2.9.1
py-moneyed==1.2
pycodestyle==2.7.0
pycparser==2.20
PyJWT==2.1.0
pyparsing==2.4.7
pyrsistent==0.18.0
python-decouple==3.4
python3-openid==3.2.0
pytz==2021.1
PyYAML==5.4.1
requests==2.25.1
requests-oauthlib==1.3.0
ruamel.yaml==0.17.10
ruamel.yaml.clib==0.2.4
six==1.16.0
sqlparse==0.4.1
swagger-spec-validator==2.7.3
toml==0.10.2
uritemplate==3.0.1
urllib3==1.26.5
